clang warning work

+++ src/ComTerp/comterp.c
@@ -739,7 +739,7 @@ boolean ComTerp::skip_arg(ComValue* topval, int& offset, int offlimit, int& tokc
 }
 
 ComValue& ComTerp::expr_top(int n) {
-  if (_pfnum+n < 0 || n>0) {
+  if (((int)_pfnum)+n < 0 || n>0) {
     return ComValue::unkval();    
   }
   else
